test: update & add tests for 6e3ad5e > move mixin tests to respective tests for each `ArgChecker` class
