+++ test/src/end_to_end_tests/modify/set_subcommand_test.dart
@@ -20,8 +20,8 @@ void main() {
   const name = 'Test File';
   const description = 'This is a test';
   const homepage = 'https://url.to.homepage';
-  const preleaseArg = '--set-prerelease=test';
-  const buildArg = '--set-build=100';
+  const prerelease = 'test';
+  const build = '100';
   const version = '8.8.8+8';
 
   setUp(() {
@@ -162,7 +162,7 @@ void main() {
 
     test('changes the prerelease in version and removes build info', () async {
       //
-      final args = [...defaultArgs, preleaseArg];
+      final args = [...defaultArgs, '--set-prerelease', prerelease];
 
       final result = await commandRunner.run(args);
 
@@ -176,7 +176,12 @@ void main() {
     });
 
     test('changes the prerelease in version and keeps build info', () async {
-      final args = [...defaultArgs, preleaseArg, '--keep-build'];
+      final args = [
+        ...defaultArgs,
+        '--set-prerelease',
+        prerelease,
+        '--keep-build',
+      ];
 
       final result = await commandRunner.run(args);
 
@@ -190,7 +195,7 @@ void main() {
     });
 
     test('changes the build and removes prerelease info', () async {
-      final args = [...defaultArgs, buildArg];
+      final args = [...defaultArgs, '--set-build', build];
 
       final result = await commandRunner.run(args);
 
@@ -204,7 +209,13 @@ void main() {
     });
 
     test('changes the build and sets prerelease info', () async {
-      final args = [...defaultArgs, preleaseArg, buildArg];
+      final args = [
+        ...defaultArgs,
+        '--set-prerelease',
+        prerelease,
+        '--set-build',
+        build,
+      ];
 
       final result = await commandRunner.run(args);
